Adding console app , more functionality checks before adding API

+++ EventManager.Service/Services/Inplementations/EventSubscriptionService.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using EventManager.Service.Exceptions;
 using EventManager.Service.Commands;
 using EventManager.Service.Models;
 using EventManager.Service.Models.Enums;
@@ -22,7 +22,7 @@ public class EventSubscriptionService
     {
         command.Validate();
         var currentEvent = await _eventRepository.GetByIdAsync(command.EventId);
-        currentEvent.EnsureIsActive();
+        if(currentEvent.EnsureIsActive()) throw new DomainException("Event is not Active or Date is invalid");
 
         if (await _eventSubscriptionRepository.Exists(command.EventId, command.UserId))
             return;
@@ -42,8 +42,9 @@ public class EventSubscriptionService
     {
         command.Validate();
         var currentEvent = await _eventRepository.GetByIdAsync(command.EventId);
-        currentEvent.EnsureIsActive();
 
-       await _eventSubscriptionRepository.RemoveSubscription(command.UserId, command.EventId);
+        if (currentEvent.EnsureIsActive()) throw new DomainException("Event is not Active or Date is invalid");
+
+        await _eventSubscriptionRepository.RemoveSubscription(command.UserId, command.EventId);
     }
 }
